Second iteration: procedural macro to derive the Planet trait
Uses a helper attribute to set the Planet's orbital period.

I thought to myself: "the usual way to auto-implement traits is to use a
derive-directive", and so I went on to do just that.
Although I've never written procedural macros before, this turned out to
be more involved than anticipated.
But with some grit, it was do-able.

Looking at other solutions, I realize that a declarative macro could
both create *and* implement the trait, as such:
```rust
boilerplate!(
    Mercury => 0.2408467,
    Earth => 1.0,
    ...
)
```

On the other hand, with a procedural macro you can slap a `Planet` onto
any struct you'd like! Maybe an Exoplanet struct that is also a Planet
but with extra fields, for example.