Third iteration: use syn::Attribute::parse_nested_meta
This meant I had to change the attribute's syntax to
`#[ident(key=value)]`
since a Meta::List, not a Meta::NameValue, is expected.

I find this strange since, to me, it seems that a MetaItem according to
the Rust reference should allow `SimplePath = Expression`.

In any case, it might be nice to also have the unit of the orbital
period in the macro.