First solution to the Hamming exercise
This uses zip_eq from the crate itertools.
I wanted a strict zip such as the one from python, but I'm not very
happy to have to deal with panicking code.

I'll probably settle for using the standard library zip, but checking
the length beforehand.
Keep this for future reference as an example of catch_unwind.